refactor: Move QuestionRepository
